requirements.txt
Django==3.0.7
djangorestframework==3.11.0
postgres==2.2.2
psycopg2>=2.7,<3.0